Add files via upload
In this folder you will find: New schema, using a bit of everyone's logic, New DB script that has worked on my laptop, still need to edit the comments on the script so it is clear and reproducible, also in the script we must add the file path to the data as it is only referred to as "******** Insert here ***** for now. You will also find the data that I used, it is in theory Astha's last clean version that I had but they may be different. I also split some files, it is explained in the script. They are also numbered in order of import.